DataManager and CreateTest changed
when a test is created it is now possible to save it to a file. tests are being saved in separate folders for each user. there is still a lot to modify: what happens when a user tries to create a test with an already existing name; are the Lists really necessary; I didnt modify IdataManager; maybe not all the methods in DataManager are needed; maybe methods/ variables in DataManager should be renamed...  but the code works so gg for now